add more fixtures

+++ src/lib.rs
@@ -5,9 +5,9 @@ pub mod dispatch;
 pub mod op;
 pub mod vm;
 
-pub fn fib(n: usize) -> f64 {
-  let mut a = 0.0;
-  let mut b = 1.0;
+pub fn fib(n: usize) -> i32 {
+  let mut a = 0;
+  let mut b = 1;
   for _ in 0..n {
     (a, b) = (b, a + b);
   }
